fix: improve write handler organization
