+++ internal/storage/istorage_engine.go
@@ -1,0 +1,7 @@
+package storage
+
+type IStorageEngine interface {
+	Put(key []byte, value []byte) error
+	Get(key []byte) ([]byte, error)
+	Delete(key []byte) error
+}
